# internal/api/openapi/server_gen.go
// Code generated by scripts/generate_openapi_go.py. DO NOT EDIT.
// Source: docs/openapi.v1.yaml

package openapi

import (
	"context"
	"encoding/json"
	"errors"
	"net/http"
	"strings"

	"cova/pkg/apiv1"
)

type SubmitBriefJobParams struct {
	XRequestId string
	IdempotencyKey string
	XTenantId string
	XProjectId string
}

type GetJobStatusParams struct {
	XRequestId string
	JobID string
	XTenantId string
	XProjectId string
}

type GetJobResultParams struct {
	XRequestId string
	JobID string
	XTenantId string
	XProjectId string
}

type GetJobDeliveriesParams struct {
	XRequestId string
	JobID string
	XTenantId string
	XProjectId string
}

type CancelJobParams struct {
	XRequestId string
	JobID string
	XTenantId string
	XProjectId string
}

type ReplayJobParams struct {
	XRequestId string
	JobID string
	XTenantId string
	XProjectId string
}

type QueryExpertParams struct {
	XRequestId string
	XTenantId string
	XProjectId string
}

type GetExpertCapabilitiesParams struct {
	XRequestId string
	XTenantId string
	XProjectId string
}

type ServerInterface interface {
	SubmitBriefJob(ctx context.Context, params SubmitBriefJobParams, req apiv1.SubmitBriefRequest) (int, any, error)
	GetJobStatus(ctx context.Context, params GetJobStatusParams) (int, any, error)
	GetJobResult(ctx context.Context, params GetJobResultParams) (int, any, error)
	GetJobDeliveries(ctx context.Context, params GetJobDeliveriesParams) (int, any, error)
	CancelJob(ctx context.Context, params CancelJobParams) (int, any, error)
	ReplayJob(ctx context.Context, params ReplayJobParams) (int, any, error)
	QueryExpert(ctx context.Context, params QueryExpertParams, req apiv1.QueryRequest) (int, any, error)
	GetExpertCapabilities(ctx context.Context, params GetExpertCapabilitiesParams) (int, any, error)
}

type Handler struct {
	impl ServerInterface
}

func NewHandler(impl ServerInterface) http.Handler {
	return &Handler{impl: impl}
}

func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	for _, route := range routes {
		if r.Method != route.method {
			continue
		}
		pathParams, ok := matchPath(route.path, r.URL.Path)
		if !ok {
			continue
		}
		h.dispatch(route.operation, w, r, pathParams)
		return
	}
	http.NotFound(w, r)
}

func (h *Handler) dispatch(operation string, w http.ResponseWriter, r *http.Request, pathParams map[string]string) {
	ctx := r.Context()
	requestID := strings.TrimSpace(r.Header.Get("X-Request-Id"))
	switch operation {
	case "submitBriefJob":
		params := SubmitBriefJobParams{
			XRequestId: strings.TrimSpace(r.Header.Get("X-Request-Id")),
			IdempotencyKey: strings.TrimSpace(r.Header.Get("Idempotency-Key")),
			XTenantId: strings.TrimSpace(r.Header.Get("X-Tenant-Id")),
			XProjectId: strings.TrimSpace(r.Header.Get("X-Project-Id")),
		}
		var req apiv1.SubmitBriefRequest
		if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
			writeProblem(w, http.StatusBadRequest, apiv1.Problem{
				Type: "https://api.cova.example.com/errors/invalid-request",
				Title: "Invalid request",
				Status: http.StatusBadRequest,
				Detail: stringPtr("request body must be valid JSON"),
				Instance: stringPtr(r.URL.Path),
				RequestId: stringPtr(requestID),
			})
			return
		}
		status, body, err := h.impl.SubmitBriefJob(ctx, params, req)
		if err != nil {
			if errors.Is(err, ErrUnimplemented) {
				writeProblem(w, http.StatusNotImplemented, apiv1.Problem{
					Type: "https://api.cova.example.com/errors/not-implemented",
					Title: "Not implemented",
					Status: http.StatusNotImplemented,
					Detail: stringPtr("operation is not implemented"),
					Instance: stringPtr(r.URL.Path),
					RequestId: stringPtr(requestID),
				})
				return
			}
			writeProblem(w, http.StatusInternalServerError, apiv1.Problem{
				Type: "https://api.cova.example.com/errors/internal-error",
				Title: "Internal error",
				Status: http.StatusInternalServerError,
				Detail: stringPtr(err.Error()),
				Instance: stringPtr(r.URL.Path),
				RequestId: stringPtr(requestID),
			})
			return
		}
		writeResponse(w, status, body)
		return
	case "getJobStatus":
		params := GetJobStatusParams{
			XRequestId: strings.TrimSpace(r.Header.Get("X-Request-Id")),
			JobID: pathParams["job_id"],
			XTenantId: strings.TrimSpace(r.Header.Get("X-Tenant-Id")),
			XProjectId: strings.TrimSpace(r.Header.Get("X-Project-Id")),
		}
		status, body, err := h.impl.GetJobStatus(ctx, params)
		if err != nil {
			if errors.Is(err, ErrUnimplemented) {
				writeProblem(w, http.StatusNotImplemented, apiv1.Problem{
					Type: "https://api.cova.example.com/errors/not-implemented",
					Title: "Not implemented",
					Status: http.StatusNotImplemented,
					Detail: stringPtr("operation is not implemented"),
					Instance: stringPtr(r.URL.Path),
					RequestId: stringPtr(requestID),
				})
				return
			}
			writeProblem(w, http.StatusInternalServerError, apiv1.Problem{
				Type: "https://api.cova.example.com/errors/internal-error",
				Title: "Internal error",
				Status: http.StatusInternalServerError,
				Detail: stringPtr(err.Error()),
				Instance: stringPtr(r.URL.Path),
				RequestId: stringPtr(requestID),
			})
			return
		}
		writeResponse(w, status, body)
		return
	case "getJobResult":
		params := GetJobResultParams{
			XRequestId: strings.TrimSpace(r.Header.Get("X-Request-Id")),
			JobID: pathParams["job_id"],
			XTenantId: strings.TrimSpace(r.Header.Get("X-Tenant-Id")),
			XProjectId: strings.TrimSpace(r.Header.Get("X-Project-Id")),
		}
		status, body, err := h.impl.GetJobResult(ctx, params)
		if err != nil {
			if errors.Is(err, ErrUnimplemented) {
				writeProblem(w, http.StatusNotImplemented, apiv1.Problem{
					Type: "https://api.cova.example.com/errors/not-implemented",
					Title: "Not implemented",
					Status: http.StatusNotImplemented,
					Detail: stringPtr("operation is not implemented"),
					Instance: stringPtr(r.URL.Path),
					RequestId: stringPtr(requestID),
				})
				return
			}
			writeProblem(w, http.StatusInternalServerError, apiv1.Problem{
				Type: "https://api.cova.example.com/errors/internal-error",
				Title: "Internal error",
				Status: http.StatusInternalServerError,
				Detail: stringPtr(err.Error()),
				Instance: stringPtr(r.URL.Path),
				RequestId: stringPtr(requestID),
			})
			return
		}
		writeResponse(w, status, body)
		return
	case "getJobDeliveries":
		params := GetJobDeliveriesParams{
			XRequestId: strings.TrimSpace(r.Header.Get("X-Request-Id")),
			JobID: pathParams["job_id"],
			XTenantId: strings.TrimSpace(r.Header.Get("X-Tenant-Id")),
			XProjectId: strings.TrimSpace(r.Header.Get("X-Project-Id")),
		}
		status, body, err := h.impl.GetJobDeliveries(ctx, params)
		if err != nil {
			if errors.Is(err, ErrUnimplemented) {
				writeProblem(w, http.StatusNotImplemented, apiv1.Problem{
					Type: "https://api.cova.example.com/errors/not-implemented",
					Title: "Not implemented",
					Status: http.StatusNotImplemented,
					Detail: stringPtr("operation is not implemented"),
					Instance: stringPtr(r.URL.Path),
					RequestId: stringPtr(requestID),
				})
				return
			}
			writeProblem(w, http.StatusInternalServerError, apiv1.Problem{
				Type: "https://api.cova.example.com/errors/internal-error",
				Title: "Internal error",
				Status: http.StatusInternalServerError,
				Detail: stringPtr(err.Error()),
				Instance: stringPtr(r.URL.Path),
				RequestId: stringPtr(requestID),
			})
			return
		}
		writeResponse(w, status, body)
		return
	case "cancelJob":
		params := CancelJobParams{
			XRequestId: strings.TrimSpace(r.Header.Get("X-Request-Id")),
			JobID: pathParams["job_id"],
			XTenantId: strings.TrimSpace(r.Header.Get("X-Tenant-Id")),
			XProjectId: strings.TrimSpace(r.Header.Get("X-Project-Id")),
		}
		status, body, err := h.impl.CancelJob(ctx, params)
		if err != nil {
			if errors.Is(err, ErrUnimplemented) {
				writeProblem(w, http.StatusNotImplemented, apiv1.Problem{
					Type: "https://api.cova.example.com/errors/not-implemented",
					Title: "Not implemented",
					Status: http.StatusNotImplemented,
					Detail: stringPtr("operation is not implemented"),
					Instance: stringPtr(r.URL.Path),
					RequestId: stringPtr(requestID),
				})
				return
			}
			writeProblem(w, http.StatusInternalServerError, apiv1.Problem{
				Type: "https://api.cova.example.com/errors/internal-error",
				Title: "Internal error",
				Status: http.StatusInternalServerError,
				Detail: stringPtr(err.Error()),
				Instance: stringPtr(r.URL.Path),
				RequestId: stringPtr(requestID),
			})
			return
		}
		writeResponse(w, status, body)
		return
	case "replayJob":
		params := ReplayJobParams{
			XRequestId: strings.TrimSpace(r.Header.Get("X-Request-Id")),
			JobID: pathParams["job_id"],
			XTenantId: strings.TrimSpace(r.Header.Get("X-Tenant-Id")),
			XProjectId: strings.TrimSpace(r.Header.Get("X-Project-Id")),
		}
		status, body, err := h.impl.ReplayJob(ctx, params)
		if err != nil {
			if errors.Is(err, ErrUnimplemented) {
				writeProblem(w, http.StatusNotImplemented, apiv1.Problem{
					Type: "https://api.cova.example.com/errors/not-implemented",
					Title: "Not implemented",
					Status: http.StatusNotImplemented,
					Detail: stringPtr("operation is not implemented"),
					Instance: stringPtr(r.URL.Path),
					RequestId: stringPtr(requestID),
				})
				return
			}
			writeProblem(w, http.StatusInternalServerError, apiv1.Problem{
				Type: "https://api.cova.example.com/errors/internal-error",
				Title: "Internal error",
				Status: http.StatusInternalServerError,
				Detail: stringPtr(err.Error()),
				Instance: stringPtr(r.URL.Path),
				RequestId: stringPtr(requestID),
			})
			return
		}
		writeResponse(w, status, body)
		return
	case "queryExpert":
		params := QueryExpertParams{
			XRequestId: strings.TrimSpace(r.Header.Get("X-Request-Id")),
			XTenantId: strings.TrimSpace(r.Header.Get("X-Tenant-Id")),
			XProjectId: strings.TrimSpace(r.Header.Get("X-Project-Id")),
		}
		var req apiv1.QueryRequest
		if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
			writeProblem(w, http.StatusBadRequest, apiv1.Problem{
				Type: "https://api.cova.example.com/errors/invalid-request",
				Title: "Invalid request",
				Status: http.StatusBadRequest,
				Detail: stringPtr("request body must be valid JSON"),
				Instance: stringPtr(r.URL.Path),
				RequestId: stringPtr(requestID),
			})
			return
		}
		status, body, err := h.impl.QueryExpert(ctx, params, req)
		if err != nil {
			if errors.Is(err, ErrUnimplemented) {
				writeProblem(w, http.StatusNotImplemented, apiv1.Problem{
					Type: "https://api.cova.example.com/errors/not-implemented",
					Title: "Not implemented",
					Status: http.StatusNotImplemented,
					Detail: stringPtr("operation is not implemented"),
					Instance: stringPtr(r.URL.Path),
					RequestId: stringPtr(requestID),
				})
				return
			}
			writeProblem(w, http.StatusInternalServerError, apiv1.Problem{
				Type: "https://api.cova.example.com/errors/internal-error",
				Title: "Internal error",
				Status: http.StatusInternalServerError,
				Detail: stringPtr(err.Error()),
				Instance: stringPtr(r.URL.Path),
				RequestId: stringPtr(requestID),
			})
			return
		}
		writeResponse(w, status, body)
		return
	case "getExpertCapabilities":
		params := GetExpertCapabilitiesParams{
			XRequestId: strings.TrimSpace(r.Header.Get("X-Request-Id")),
			XTenantId: strings.TrimSpace(r.Header.Get("X-Tenant-Id")),
			XProjectId: strings.TrimSpace(r.Header.Get("X-Project-Id")),
		}
		status, body, err := h.impl.GetExpertCapabilities(ctx, params)
		if err != nil {
			if errors.Is(err, ErrUnimplemented) {
				writeProblem(w, http.StatusNotImplemented, apiv1.Problem{
					Type: "https://api.cova.example.com/errors/not-implemented",
					Title: "Not implemented",
					Status: http.StatusNotImplemented,
					Detail: stringPtr("operation is not implemented"),
					Instance: stringPtr(r.URL.Path),
					RequestId: stringPtr(requestID),
				})
				return
			}
			writeProblem(w, http.StatusInternalServerError, apiv1.Problem{
				Type: "https://api.cova.example.com/errors/internal-error",
				Title: "Internal error",
				Status: http.StatusInternalServerError,
				Detail: stringPtr(err.Error()),
				Instance: stringPtr(r.URL.Path),
				RequestId: stringPtr(requestID),
			})
			return
		}
		writeResponse(w, status, body)
		return
	default:
		http.NotFound(w, r)
	}
}

var ErrUnimplemented = errors.New("operation not implemented")

type route struct {
	method    string
	path      string
	operation string
}

var routes = []route{
	{method: "POST", path: "/v1/assistant/jobs", operation: "submitBriefJob"},
	{method: "GET", path: "/v1/assistant/jobs/{job_id}", operation: "getJobStatus"},
	{method: "GET", path: "/v1/assistant/jobs/{job_id}/result", operation: "getJobResult"},
	{method: "GET", path: "/v1/assistant/jobs/{job_id}/deliveries", operation: "getJobDeliveries"},
	{method: "POST", path: "/v1/assistant/jobs/{job_id}/cancel", operation: "cancelJob"},
	{method: "POST", path: "/v1/assistant/jobs/{job_id}/replay", operation: "replayJob"},
	{method: "POST", path: "/v1/assistant/query", operation: "queryExpert"},
	{method: "GET", path: "/v1/assistant/experts", operation: "getExpertCapabilities"},
}

func matchPath(pattern, actual string) (map[string]string, bool) {
	patternParts := splitPath(pattern)
	actualParts := splitPath(actual)
	if len(patternParts) != len(actualParts) {
		return nil, false
	}
	params := make(map[string]string)
	for i := range patternParts {
		p := patternParts[i]
		a := actualParts[i]
		if strings.HasPrefix(p, "{") && strings.HasSuffix(p, "}") {
			key := strings.TrimSuffix(strings.TrimPrefix(p, "{"), "}")
			params[key] = a
			continue
		}
		if p != a {
			return nil, false
		}
	}
	return params, true
}

func splitPath(path string) []string {
	trimmed := strings.Trim(path, "/")
	if trimmed == "" {
		return []string{}
	}
	return strings.Split(trimmed, "/")
}

func stringPtr(v string) *string {
	return &v
}

func writeProblem(w http.ResponseWriter, status int, p apiv1.Problem) {
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(p)
}

func writeResponse(w http.ResponseWriter, status int, body any) {
	if body == nil {
		w.WriteHeader(status)
		return
	}
	if status >= 400 {
		w.Header().Set("Content-Type", "application/problem+json")
	} else {
		w.Header().Set("Content-Type", "application/json")
	}
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(body)
}